+++ app/helpers/application_helper.rb
@@ -1,2 +1,4 @@
 module ApplicationHelper
+    # Pagination module
+    include Pagy::Frontend
 end